+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Games:
 - [Savage 2](https://savage2.net/) support.
 - [Rising World](https://store.steampowered.com/app/324080/Rising_World/) support.
 - [ATLAS](https://store.steampowered.com/app/834910/ATLAS/) support.
+- [America's Army: Proving Grounds](https://store.steampowered.com/app/203290/Americas_Army_Proving_Grounds/) support.
 - Added a valve protocol query example.
 - Made all of Just Cause 2: Multiplayer Response and Player fields public.